Fix type conversion related warnings for common/base/task*
Depends-On: I2b4ba7fbeccd4128e8b23d408fa999d15b245fcb
Partial-Bug: #1800894

+++ src/vnsw/agent/cmn/agent.cc
@@ -1013,7 +1013,7 @@ boost::shared_ptr<AgentXmppChannel> Agent::controller_xmpp_channel_ref(uint8_t i
 
 void Agent::TaskTrace(const char *file_name, uint32_t line_no,
                       const Task *task, const char *description,
-                      uint32_t delay) {
+                      uint64_t delay) {
     TaskTrace::TraceMsg(TaskTraceBuf, file_name, line_no,
                         task->GetTaskId(), task->GetTaskInstance(),
                         description, delay, task->Description());
